Fix RSpec
