+++ config/initializers/budgets_workflows.rb
@@ -1,3 +1,3 @@
 Rails.application.config.to_prepare do
-  require "./lib/decidim/budgets/workflows"
+  require "./lib/extensions/decidim/budgets/workflows"
 end
