--- lib/decidim/budgets/workflows.rb
@@ -1,3 +1,0 @@
-require "decidim/budgets/workflows/monterrey"
-
-Decidim::Budgets.workflows[:monterrey] = Decidim::Budgets::Workflows::Monterrey
