+++ lib/extensions/decidim/budgets/workflows.rb
@@ -1,0 +1,3 @@
+require "extensions/decidim/budgets/workflows/monterrey"
+
+::Decidim::Budgets.workflows[:monterrey] = Extensions::Decidim::Budgets::Workflows::Monterrey
